requirements.txt
# InfraSight-AI Requirements
# Python 3.10+

# Core
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
neo4j==5.14.0
pgvector==0.2.4

# AI/ML
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
torch>=2.0.0

# Data Processing
pandas==2.1.4
numpy>=1.24.0

# Web UI
gradio==4.12.0

# Data Pipeline
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.8.0

# Utilities
requests==2.31.0
pyyaml==6.0.1
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
mypy==1.7.1
